rues_api/requirements.txt
Flask
SQLAlchemy
psycopg2-binary
python-dotenv
Alembic
pytest
requests
